+++ libs/aries-basic-controller/aries_basic_controller/controllers/issuer.py
@@ -1,7 +1,7 @@
-from .base_controller import BaseController
-from aiohttp import ClientSession, ClientResponseError
+from .base import BaseController
+from aiohttp import ClientSession
 import logging
-from.utils import extract_did, get_schema_details
+from ..helpers.utils import extract_did, get_schema_details
 logger = logging.getLogger("aries_controller.issuer")
 
 CRED_PREVIEW = "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/issue-credential/1.0/credential-preview"
