+++ libs/aries-basic-controller/aries_basic_controller/controllers/wallet.py
@@ -1,4 +1,4 @@
-from .base_controller import BaseController
+from .base import BaseController
 from aiohttp import ClientSession
 import logging
